move Igenricrepo to application layer and customize genric repo

+++ Workspace.Application/Interfaces/IGenricRepo.cs
@@ -9,7 +9,7 @@ namespace Workspace.Infrastructure.Repositories.Interfaces
     public interface IGenricRepo<T> where T : class
     {
         Task<IEnumerable<T>> getAll();
-        Task<T> getById(int id);
+        Task<T> getById(long id);
         Task add(T entity);
         void update(T entity);
         void delete(T entity);
